+++ recipes/install-logstash.rb
@@ -5,6 +5,8 @@
 
 elk_info = get_elk_info
 
+stack_name = stack_shortname
+es_host = node[:opsworks][:instance][:private_ip]
 logstash_major_version = elk_info['logstash_major_version']
 logstash_repo_uri = elk_info['logstash_repo_uri']
 
@@ -19,10 +21,30 @@ include_recipe "mh-opsworks-recipes::update-package-repo"
 pin_package("logstash", "#{logstash_major_version}.*")
 install_package("logstash")
 
+service 'logstash' do
+  action :enable
+  supports :restart => true
+end
+
 cookbook_file "/etc/default/logstash" do
   source "logstash-default"
   owner 'root'
   group 'root'
   mode '644'
+end
+
+template '/etc/logstash/conf.d/logstash.conf' do
+  source 'logstash.conf.erb'
+  user 'root'
+  group 'root'
+  mode '644'
+  variables({
+    tcp_port: elk_info['logstash_tcp_port'],
+    sqs_queue_name: "#{stack_name}-user-actions",
+    stdout_output: elk_info['logstash_stdout_output'],
+    elasticsearch_index_prefix: elk_info['es_index_prefix'],
+    elasticsearch_host: es_host
+  })
+  notifies :restart, 'service[logstash]', :immediately
 end